Create Tic_Tac_Toe_with_Computer.java
This is a java program that serves as a tic-tac-toe game between user and computer. Here user just need to enter the coordinates of the place where-ever he/she wishes to put his/her mark. Computer automatically takes its turn and the game goes on.

Please don't forget to change the class name from Main if it throws error.